+++ modules/core/src/main/java/com/ae/apps/lib/review/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * An implementation for Google Plat In-App Reviews API
+ *
+ * @since 4.1
+ * @author midhunhk
+ */
+package com.ae.apps.lib.review;